api/requirements.txt
africastalking==1.2.7
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
crispy-bootstrap4==2024.1
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==4.2.11
django-allauth==0.62.1
django-crispy-forms==2.1
django-oauth-toolkit==2.3.0
django-oauth2-provider==0.2.6.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
exceptiongroup==1.2.1
h11==0.14.0
idna==3.7
iniconfig==2.0.0
jwcrypto==1.5.6
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
PySocks==1.7.1
pytest==8.2.0
pytest-django==4.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
schema==0.7.7
selenium==4.20.0
shortuuid==1.0.13
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sortedcontainers==2.4.0
sqlparse==0.5.0
tomli==2.0.1
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wsproto==1.2.0
